refactor(rust): extract SchemaAnalysis module
